Big update of scripts: features wrapped in functions; new scripts added
A lot of features were wrapped in new functions so that they can be imported in other new scripts when needed. Also, new functionalities were added such as pdf converting feature and the first part of new feature for extracting objects from jpg files with the help of OpenCV library.